Seprate borderpx for floating windows
Useful for having only floating windows have border which I like. It saves space
when having multiple non-floating windows open, allows video player have no
scaling when it's the only non-floating window and the video matches the screen
size, and allows to have floating windows on top of the video.